requirements.txt
torch
pillow
numpy
python-dotenv
google-generativeai
google-genai
